+++ saw-remote-api/src/SAWServer/Data/SetupValue.hs
@@ -50,7 +50,7 @@ instance (FromJSON ty, FromJSON cryptolExpr) => FromJSON (CrucibleSetupVal ty cr
           TagNullValue -> pure NullValue
           TagCryptol -> CryptolExpr <$> o .: "expression"
           TagArrayValue -> ArrayValue <$> o .:? "element type" <*> o .: "elements"
-          TagStructValue -> StructValue <$> o .: "elements"
+          TagStructValue -> StructValue <$> o .:? "MIR ADT server name" <*> o .: "elements"
           TagTupleValue -> TupleValue <$> o .: "elements"
           TagFieldLValue -> FieldLValue <$> o .: "base" <*> o .: "field"
           TagCastLValue -> CastLValue <$> o .: "base" <*> o .: "type"